Fixing the ubuntu iso download locatino (#6)
The ubuntu download location was renamed after 20.04.2 was release per #5 .

This should pin this repo to 20.04.1 moving forward. Once 20.04.3 is released we can rev this repo to 20.04.2 using this archive repo.

I'm not sure of a better way than being (N-1) to solve this since there is no static naming that remains the same. 

Maybe I could pull the latest release name dynamically somehow?